# [perf] optimize allocations and UTF-8 decoding of strings for otlp_bytes -> views -> OTAP