--- a/rust/otap-dataflow/crates/pdata-views/src/otlp/bytes/logs.rs
+++ b/rust/otap-dataflow/crates/pdata-views/src/otlp/bytes/logs.rs
@@ -327,10 +327,8 @@ impl ResourceLogsView for RawResourceLogs<'_> {
 
     #[inline]
     fn schema_url(&self) -> Option<crate::views::common::Str<'_>> {
-        let slice = self
-            .byte_parser
-            .advance_to_find_field(RESOURCE_LOGS_SCHEMA_URL)?;
-        std::str::from_utf8(slice).ok()
+        self.byte_parser
+            .advance_to_find_field(RESOURCE_LOGS_SCHEMA_URL)
     }
 
     #[inline]
@@ -372,10 +370,8 @@ impl ScopeLogsView for RawScopeLogs<'_> {
 
     #[inline]
     fn schema_url(&self) -> Option<crate::views::common::Str<'_>> {
-        let slice = self
-            .byte_parser
-            .advance_to_find_field(SCOPE_LOGS_SCHEMA_URL)?;
-        std::str::from_utf8(slice).ok()
+        self.byte_parser
+            .advance_to_find_field(SCOPE_LOGS_SCHEMA_URL)
     }
 
     #[inline]
@@ -452,10 +448,8 @@ impl LogRecordView for RawLogRecord<'_> {
 
     #[inline]
     fn severity_text(&self) -> Option<crate::views::common::Str<'_>> {
-        let slice = self
-            .bytes_parser
-            .advance_to_find_field(LOG_RECORD_SEVERITY_TEXT)?;
-        std::str::from_utf8(slice).ok()
+        self.bytes_parser
+            .advance_to_find_field(LOG_RECORD_SEVERITY_TEXT)
     }
 
     #[inline]
@@ -483,9 +477,7 @@ impl LogRecordView for RawLogRecord<'_> {
 
     #[inline]
     fn event_name(&self) -> Option<crate::views::common::Str<'_>> {
-        let slice = self
-            .bytes_parser
-            .advance_to_find_field(LOG_RECORD_EVENT_NAME)?;
-        std::str::from_utf8(slice).ok()
+        self.bytes_parser
+            .advance_to_find_field(LOG_RECORD_EVENT_NAME)
     }
 }