Closes #751 

Use BinaryArray builders where possible to lazily validate the UTF-8 Decoding when converting from OTLP Bytes -> OTAP. This is a nice optimization because we don't validate strings in the views, multiple times when inserting them in dictionary values. Instead, the UTF-8 Validation happens when we convert the Binary array to a String array.

~5.5% perf improvement encoding OTLP bytes -> OTAP
```
encode_otap_logs_using_views/proto_bytes->views->OTAP/resources=1,scopes=1,logs=1000
                        time:   [542.11 µs 542.41 µs 542.73 µs]
                        change: [−5.5094% −5.4312% −5.3581%] (p = 0.00 < 0.05)

encode_otap_logs_using_views/proto_bytes->views->OTAP/resources=1,scopes=1,logs=5000
                        time:   [2.6414 ms 2.6429 ms 2.6443 ms]
                        change: [−5.4051% −5.3303% −5.2524%] (p = 0.00 < 0.05)

encode_otap_logs_using_views/proto_bytes->views->OTAP/resources=1,scopes=1,logs=10000
                        time:   [5.2606 ms 5.2647 ms 5.2691 ms]
                        change: [−5.1650% −5.0651% −4.9688%] (p = 0.00 < 0.05)
```

Some improvement building OTAP from syslog as well
```
arrow_batch_creation/rfc3164_arrow_batch_100_msgs
                        time:   [73.360 µs 73.489 µs 73.627 µs]
                        change: [−2.0723% −1.7579% −1.4337%] (p = 0.00 < 0.05)
                        Performance has improved.

arrow_batch_creation/rfc5424_arrow_batch_100_msgs
                        time:   [42.795 µs 42.850 µs 42.904 µs]
                        change: [−0.1449% +0.0844% +0.2942%] (p = 0.45 > 0.05)
                        No change in performance detected.

arrow_batch_creation/cef_arrow_batch_100_msgs
                        time:   [35.831 µs 35.957 µs 36.093 µs]
                        change: [−5.6171% −5.2694% −4.9296%] (p = 0.00 < 0.05)
                        Performance has improved.
```